refactors parser to add support for dependency injection and removes paid
