+++ src/parser/halt.factory.js
@@ -1,0 +1,9 @@
+module.exports = HALT => {
+  return command => {
+    const event = command.slice(HALT.length + 1);
+    return {
+      type: HALT,
+      event
+    };
+  };
+};
