+++ src/parser/halt.js
@@ -1,11 +1,2 @@
 const { HALT } = require('../commands');
-
-function parseHaltCommand(command) {
-  const event = command.slice(HALT.length + 1);
-  return {
-    type: HALT,
-    event
-  };
-}
-
-module.exports = { parseHaltCommand };
+module.exports = { parseHaltCommand: require('./halt.factory')(HALT) };
