+++ src/parser/move.js
@@ -1,14 +1,4 @@
 const { split } = require('./helpers');
 const { MOVE } = require('../commands');
 
-function parseMoveCommand(command) {
-  const withoutPrefix = command.slice(MOVE.length + 1);
-  const [event, to] = split(withoutPrefix, ' to ');
-  return {
-    type: MOVE,
-    event,
-    to
-  };
-}
-
-module.exports = { parseMoveCommand };
+module.exports = { parseMoveCommand: require('./move.factory')(split, MOVE) };
