--- src/parser/paid.js
@@ -1,16 +1,0 @@
-const { splitBy } = require('./helpers');
-const { PAID } = require('../commands');
-
-function parsePaidCommand(command) {
-  const withoutInfix = command.replace(' paid ', ' ');
-  const [who, to, some] = splitBy(withoutInfix, [' ', ' ']);
-
-  return {
-    type: PAID,
-    who,
-    to,
-    some
-  };
-}
-
-module.exports = { parsePaidCommand };
