+++ src/parser/resume.factory.js
@@ -1,0 +1,9 @@
+module.exports = RESUME => {
+  return command => {
+    const event = command.slice(RESUME.length + 1);
+    return {
+      type: RESUME,
+      event
+    };
+  };
+};
