+++ src/parser/resume.js
@@ -1,11 +1,2 @@
 const { RESUME } = require('../commands');
-
-function parseResumeCommand(command) {
-  const event = command.slice(RESUME.length + 1);
-  return {
-    type: RESUME,
-    event
-  };
-}
-
-module.exports = { parseResumeCommand };
+module.exports = { parseResumeCommand: require('./resume.factory')(RESUME) };
