+++ src/parser/skip.js
@@ -1,16 +1,4 @@
 const { splitBy } = require('./helpers');
 const { SKIP } = require('../commands');
 
-function parseSkipCommand(command) {
-  const withoutInfix = command.replace(' will skip ', ' ');
-  const [username, name, period] = splitBy(withoutInfix, [' ', 'for']);
-
-  return {
-    type: SKIP,
-    for: period,
-    username,
-    name
-  };
-}
-
-module.exports = { parseSkipCommand };
+module.exports = { parseSkipCommand: require('./skip.factory')(splitBy, SKIP) };
