+++ src/parser/status.factory.js
@@ -1,0 +1,5 @@
+module.exports = STATUS => {
+  return command => {
+    return { type: STATUS };
+  };
+};
