+++ src/parser/terminate.factory.js
@@ -1,0 +1,9 @@
+module.exports = (split, TERMINATE) => {
+  return command => {
+    const [type, event] = split(command);
+    return {
+      type: TERMINATE,
+      event
+    };
+  };
+};
