+++ src/parser/terminate.js
@@ -1,12 +1,6 @@
 const { split } = require('./helpers');
 const { TERMINATE } = require('../commands');
 
-function parseTerminateCommand(command) {
-  const [type, event] = split(command);
-  return {
-    type: TERMINATE,
-    event
-  };
-}
-
-module.exports = { parseTerminateCommand };
+module.exports = {
+  parseTerminateCommand: require('./terminate.factory')(split, TERMINATE)
+};
